Added health kits and a spawner

+++ Assets/Scripts/Mines/StandardMine.cs
@@ -17,8 +17,8 @@ public class StandardMine : NetworkBehaviour
        if(!health) return;
        health.TakeDamage(25);
         
-       int xPosition = Random.Range(-4, 4);
-       int yPosition = Random.Range(-2, 2);
+       float xPosition = Random.Range(-4, 4);
+       float yPosition = Random.Range(-2, 2);
        
        GameObject newMine = Instantiate(minePrefab, new Vector3(xPosition, yPosition, 0), Quaternion.identity);
        NetworkObject no = newMine.GetComponent<NetworkObject>();
